api/requirements.txt
# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.10.6
pydantic-settings==2.7.1

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# Redis & Celery
redis==5.0.1
celery==5.3.6

# Storage Drivers (Phase 2)
boto3==1.34.20
aioboto3==12.3.0
dropbox==12.0.2
aiofiles==23.2.1

# Image Processing (Phase 2)
Pillow==10.2.0

# Security (Phase 2)
cryptography==42.0.0

# Utilities
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Testing (dev dependencies)
pytest==7.4.3
pytest-cov==4.1.0
httpx==0.26.0
moto[s3]==5.0.0

# Linting & Formatting (dev dependencies)
ruff==0.1.11
black==23.12.1
